Добавил модули, разделил имеющийся код на модули.
