+++ js/constants.js
@@ -1,0 +1,35 @@
+// модуль для констант
+const RANDOM_MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+const RANDOM_NAMES = [
+  'Никита',
+  'Юлия',
+  'Паха',
+  'Дмитрий',
+  'Таньяна',
+  'Валерий',
+  'Валентина',
+];
+
+const RANDOM_DESCRIPTIONS = [
+  'фотография красивая',
+  'фотография необычная',
+  'фотография яркая',
+  'фотография живая',
+  'фотография солнечная',
+  'фотография насыщенная',
+];
+
+const GENERATED_QUANTITY_OBJECTS = 25;
+
+const AvatarName = {MIN: 1, MAX: 6};
+
+const RandomLike = {MIN: 15, MAX: 200};
+
+export {RANDOM_DESCRIPTIONS, RANDOM_NAMES, RANDOM_MESSAGES, GENERATED_QUANTITY_OBJECTS, AvatarName, RandomLike};
