--- js/random-data.js
@@ -1,1 +1,0 @@
-// модуль для генерации случайных данных ( случайное имя, случайный комментарий)
